# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wren (wren-external)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users/{handle}": {
      "get": {
        "tags": [
          "listing"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_users__handle__get",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Profile": {
        "properties": {
          "handle": {
            "type": "string",
            "title": "Handle"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "roadmaps": {
            "items": {
              "$ref": "#/components/schemas/RoadmapCard"
            },
            "type": "array",
            "title": "Roadmaps"
          }
        },
        "type": "object",
        "required": [
          "handle",
          "display_name"
        ],
        "title": "Profile",
        "description": "The ``GET /users/{handle}`` body: a user's public profile (spec section 02\nUS-ACCT-03).\n\n``roadmaps`` is only that user's **published, public** roadmaps; drafts,\nprivate, and archived roadmaps never appear, and following is never exposed\n(no social graph). Public and viewer-agnostic: the same body regardless of who\n(if anyone) is signed in."
      },
...
      "RoadmapCard": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/RoadmapStatus"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          },
          "subject_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subject Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "visibility"
        ],
        "title": "RoadmapCard",
        "description": "A roadmap summarized for a list/grid card.\n\n``status`` drives the Draft/Published/Archived badge and ``visibility`` the\nlock/globe badge. Content (sections, items, progress)\nis never inlined: a card links to the full roadmap view for that."
      },
...
      "RoadmapStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ],
        "title": "RoadmapStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Visibility": {
        "type": "string",
        "enum": [
          "public",
          "private"
        ],
        "title": "Visibility"
      },
...
    }
  }
}